refactor: Service 테스트 없애고 Dao 테스트로 변경
